move classes to a new file and update the code
